requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
google-genai==1.53.0
groq==0.37.1